Refactor: Update the client address information
