+++ src/main/java/br/com/client/micro/domain/Client.java
@@ -1,5 +1,7 @@
 package br.com.client.micro.domain;
 
+import br.com.client.micro.domain.complements.Address;
+import br.com.client.micro.domain.complements.Phone;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.Id;
