+++ src/main/java/br/com/client/micro/domain/complements/Address.java
@@ -1,5 +1,6 @@
-package br.com.client.micro.domain;
+package br.com.client.micro.domain.complements;
 
+import br.com.client.micro.domain.enums.UF;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -7,6 +8,9 @@ import lombok.Getter;
 @Getter
 public class Address {
     private String cep;
+    private UF state;
+    private String city;
+    private String street;
     private String number;
     private String complement;
 }
