+++ src/main/java/br/com/client/micro/domain/complements/Phone.java
@@ -1,4 +1,4 @@
-package br.com.client.micro.domain;
+package br.com.client.micro.domain.complements;
 
 import lombok.Builder;
 import lombok.Getter;
