+++ src/main/java/br/com/client/micro/domain/enums/Status.java
@@ -1,4 +1,4 @@
-package br.com.client.micro.domain;
+package br.com.client.micro.domain.enums;
 
 public enum Status {
     CREATED,
