+++ src/main/java/br/com/client/micro/domain/enums/UF.java
@@ -1,0 +1,7 @@
+package br.com.client.micro.domain.enums;
+
+public enum UF {
+    AC, AL, AP, AM, BA, CE, DF, ES, GO,
+    MA, MT, MS, MG, PA, PB, PR, PE, PI,
+    RJ, RN, RS, RO, RR, SC, SP, SE, TO,
+}
