+++ src/main/java/br/com/client/micro/event/dto/DeliveryEventDto.java
@@ -1,6 +1,6 @@
 package br.com.client.micro.event.dto;
 
-import br.com.client.micro.domain.Status;
+import br.com.client.micro.domain.enums.Status;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
